update to check the directory first

+++ admin/delete_file.php
@@ -7,7 +7,12 @@ if (!isset($_SERVER['HTTP_X_REQUESTED_WITH'])){
     die(0);
 }
 
-if(isset($_POST['file']))
-    die(unlink($_POST['file']));
+if(isset($_POST['file'])) {
+    $dir = explode('/',$_POST['file']);
+    if($dir[1] == 'uploads')
+        die(unlink($_POST['file']));
+    else
+        die(0);
+}
  else
     die(0);
